requirements_library.txt
streamlit
langchain
langchain-google-genai
PyPDF2
faiss-cpu
langchain-community
python-dotenv